Move `pkg/resourcemanager/predicate/has_finalizer.go` to `pkg/controllerutils/predicate/has_finalizer.go`
